add command to run a command across all nodes of a given type
